Add DesignExplorer ToolStripMenu
