+++ Tunny/UI/DEStudyNameSelector/DEStudyNameSelector.resx
@@ -117,50 +117,4 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <assembly alias="System.Windows.Forms" name="System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089" />
-  <data name="BoneFish" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>BoneFish.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="ConstructFishAttribute" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>ConstructFishAttribute.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="ConstructFishEgg" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>ConstructFishEgg.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="DeconstructFish" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>DeconstructFish.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="DeconstructFishAttribute" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>DeconstructFishAttribute.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="FishMarket" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>FishMarket.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="FishPrintByCapture" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>FishPrintByCapture.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="FishPrintByPath" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>FishPrintByPath.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="optuna_dashboard" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>optuna-dashboard.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="ParamFishAttribute" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>ParamFishAttribute.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="ParamFishEgg" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>ParamFishEgg.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="ParamFishIcon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>ParamFishIcon.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="ParamFishPrintIcon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>ParamFishPrintIcon.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="TTDesignExplorer" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>DesignExplorer.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
-  <data name="TunnyIcon" type="System.Resources.ResXFileRef, System.Windows.Forms">
-    <value>TunnyIcon.png;System.Drawing.Bitmap, System.Drawing, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a</value>
-  </data>
 </root>
