+++ Tunny/UI/PythonInstallDialog/PythonInstallDialog.resx
