+++ Tunny/UI/TunnyUI/OptimizationWindow.cs
