+++ Tunny/UI/TunnyUI/OptimizationWindow.resx
